Moved the types to the shared types file
