+++ server/src/utils/concurrencyLimiter.ts
@@ -1,3 +1,5 @@
+import type { ConcurrencyLimiterOpts, ConcurrencyLimiter } from '#src/utils/types.ts'
+
 /** A generic concurrency gating device. `createConcurrencyLimiter` returns a
  * limiter that runs at most `maxConcurrent` async tasks at once; further tasks
  * wait in a bounded FIFO queue, and once that queue is full the limiter sheds
@@ -7,30 +9,9 @@
  *
  * Each limiter is an independent instance holding all of its state in a
  * closure, so one process can gate several distinct resources with separate
- * caps. This was lifted verbatim from the volcano render gate that previously
- * lived in renderVolcano.ts; see utils/cacheOrRecompute.ts for the sibling
- * pattern that gates concurrent compute jobs. */
-
-export type ConcurrencyLimiterOpts = {
-	/** Max tasks allowed to run at the same time. */
-	maxConcurrent: number
-
-	/** Max tasks allowed to wait for a slot. Past this, `run` rejects
-	 * immediately with the busy error instead of enqueuing. */
-	maxQueued: number
-
-	/** Optional factory for the error thrown when the wait-queue is full.
-	 * Defaults to a generic 429 (`code: 'POOL_BUSY'`). Supply your own to
-	 * carry a caller-specific code/message (e.g. volcano's `RENDER_BUSY`). */
-	makeBusyError?: () => Error
-}
-
-export type ConcurrencyLimiter = {
-	/** Acquire a slot (waiting or rejecting per the caps), run `fn`, and
-	 * release the slot in a `finally` so it can never leak — even if `fn`
-	 * throws. Resolves/rejects with `fn`'s own result. */
-	run<T>(fn: () => Promise<T>): Promise<T>
-}
+ * caps. See utils/cacheOrRecompute.ts for the sibling
+ * pattern that gates concurrent compute jobs. The `ConcurrencyLimiterOpts` and
+ * `ConcurrencyLimiter` types live in utils/types.ts alongside cacheOrRecompute's. */
 
 /** Generic 429 so it reads as "retry later", not a client error in the
  * request shape. Mirrors the busy-error shape in utils/cacheOrRecompute.ts. */
