+++ server/src/utils/types.ts
@@ -24,3 +24,24 @@ export type CacheOrRecomputeResult<TResult> = {
 	cacheId: string
 	cacheFilePath: string
 }
+
+export type ConcurrencyLimiterOpts = {
+	/** Max tasks allowed to run at the same time. */
+	maxConcurrent: number
+
+	/** Max tasks allowed to wait for a slot. Past this, `run` rejects
+	 * immediately with the busy error instead of enqueuing. */
+	maxQueued: number
+
+	/** Optional factory for the error thrown when the wait-queue is full.
+	 * Defaults to a generic 429 (`code: 'POOL_BUSY'`). Supply your own to
+	 * carry a caller-specific code/message (e.g. volcano's `RENDER_BUSY`). */
+	makeBusyError?: () => Error
+}
+
+export type ConcurrencyLimiter = {
+	/** Acquire a slot (waiting or rejecting per the caps), run `fn`, and
+	 * release the slot in a `finally` so it can never leak — even if `fn`
+	 * throws. Resolves/rejects with `fn`'s own result. */
+	run<T>(fn: () => Promise<T>): Promise<T>
+}
